Create 300x300x5 grid example
# Summary
This example demonstrate a decent chunk rendering with stable performance on my hardware.

It uses instancing to perform this (so one drawcall for the entire chunk, allocating a buffer for the model).

Additionally, it also redefines the stack to refer to a vertical stack, as it was more intituive when setting the number of voxels.